+++ CmdLine.hh
@@ -298,11 +298,13 @@ class CmdLine {
   CmdLine & help(const std::string & help_str);
 
   /// start a section of the help
-  void start_section(const std::string & section_name) {
+  void start_section(const std::string & section_name, const std::string & description = "") {
     __current_section = section_name;
     __current_subsection = "";
+    if (description != "") __section_descriptions[__section_key(section_name)] = description;
   }
-  void section(const std::string & section_name) {start_section(section_name);}
+  void section(const std::string & section_name, const std::string & description = "") {
+    start_section(section_name, description);}
 
   /// end a section of the help
   void end_section() {__current_section = ""; __current_subsection = "";}
@@ -315,8 +317,10 @@ class CmdLine {
 
 
   /// start a subsection of the help
-  void start_subsection(const std::string & subsection_name);
-  void subsection(const std::string & subsection_name) {start_subsection(subsection_name);}
+  void start_subsection(const std::string & subsection_name, const std::string & description = "");
+  void subsection(const std::string & subsection_name, const std::string & description = "") {
+    start_subsection(subsection_name, description);
+  }
 
   /// end a subsection of the help
   void end_subsection() {__current_subsection = "";}
@@ -538,6 +542,13 @@ class CmdLine {
   std::string __current_section = "";
   std::string __current_subsection = "";
 
+  /// map of description for each section, organised according to the key
+  /// as produced by __section_key
+  std::map<std::string,std::string> __section_descriptions;
+  inline static std::string __section_key(const std::string & section_name, const std::string & subsection_name = "") {
+    return "SEC:"+section_name + (subsection_name == "" ? "" : "-SUBSEC:"+subsection_name);
+  }
+
   /// default option to tell CmdLine to read arguments 
   /// from a file
   static std::string _default_argfile_option;
@@ -547,6 +558,7 @@ class CmdLine {
   struct OptSection {
     OptSection(const std::string & name_in, int level_in) : name(name_in), level(level_in) {}
     std::string name;
+    std::string section_key;
     int level;
     std::vector<const OptionHelp *> options;
   };