+++ unit-tests.cc
@@ -165,14 +165,20 @@ int main() {
     CHECK_PASS(cmd, "-f .false. -i 2", make_tuple(2,false));
   }
 
+  // check the default constructor compiles. We will assign
+  // to it below to check that that works too.
+  CmdLine::Result<double> double_result;
+
   //---------------------------------------------------------------------------
   // verify cases that take initialiser lists and vectors of options
-  auto cmd = [](CmdLine & cmdline){
+  auto cmd = [&double_result](CmdLine & cmdline){
     vector<string> opts_d = {"-d","--double"};
     auto uu = cmdline.optional_value<int>({"-u","--uu"});
     double u = uu.present() ? uu.value() : 3;
+
+    double_result = cmdline.value<double>(opts_d, 1.4);
     return make_tuple(      
-      cmdline.value<double>(opts_d, 1.4),
+      double_result,
       cmdline.value<int>({"-i","--int"}),
       u
     );
@@ -182,6 +188,8 @@ int main() {
   CHECK_PASS(cmd, "--int 2",              make_tuple(1.4, 2, 3));
   CHECK_PASS(cmd, "--int 2 --uu 6",       make_tuple(1.4, 2, 6));
   CHECK_FAIL(cmd, "");
+
+
 
   cout << "All " << n_checks << " checks passed" << endl;
   return 0;